added delta_v_reward tests, removed delta_v modification, added bug fix to ensure delta_v_reward always negative
