+++ safe_autonomy_sims/gym/docking/utils.py
@@ -81,7 +81,7 @@ def delta_v(state: dict, prev_state: dict):
     """
     v = np.linalg.norm(state["deputy"][3:6])
     prev_v = np.linalg.norm(prev_state["deputy"][3:6])
-    return np.abs(v - prev_v)
+    return v - prev_v
 
 
 def v_limit(
