+++ test/unit_tests/safe_autonomy_sims/gym/docking/reward/delta_v_reward_test_cases.yml
@@ -5,7 +5,7 @@
   state:
     chief: np.array([0,0,0,0,0,0])
     deputy: np.array([10,10,10,2,2,2])
-  expected_value: 1.7320508075688772
+  expected_value: -0.14433756729740643
 - ID: negative velocity change
   prev_state: 
     chief: np.array([0,0,0,0,0,0])
@@ -13,4 +13,4 @@
   state:
     chief: np.array([0,0,0,0,0,0])
     deputy: np.array([10,10,10,0.5,0.5,0.5])
-  expected_value: -0.8660254037844386
+  expected_value: -0.07216878364870322
