+++ test/unit_tests/safe_autonomy_sims/gym/inspection/reward/delta_v_reward_test_cases.yml
@@ -1,8 +1,8 @@
 - ID: positive velocity change
   prev_v: np.array([1,1,1])
   v: np.array([2,2,2])
-  expected_value: 1.7320508075688772
+  expected_value: -0.14433756729740643
 - ID: negative velocity change
   prev_v: np.array([1,1,1])
   v: np.array([0.5,0.5,0.5])
-  expected_value: -0.8660254037844386
+  expected_value: -0.07216878364870322
